+++ app/src/main/java/nextstep/payments/screens/card/mapper/CardMapper.kt
@@ -1,8 +1,6 @@
 package nextstep.payments.screens.card.mapper
 
 import nextstep.payments.domain.Card
-import nextstep.payments.screens.card.toDomain
-import nextstep.payments.screens.card.toState
 import nextstep.payments.screens.card.update.CardState
 
 fun CardState.toDomain(): Card? {
